Add main.go
Change fop package same as file name.
Move error into a new file.
