+++ main.go
@@ -1,0 +1,10 @@
+package main
+
+import (
+	. "github.com/kiedveian/JDExam/fops"
+	"os"
+)
+
+func main() {
+	RunFops(os.Args[1:])
+}
